requirements.txt
altair
jupyterlab
ipywidgets
matplotlib
seaborn
numpy
pandas
jupyterlab-dash
scikit-learn
statsmodels
xgboost
python-dotenv
pytest
spotipy
tensorflow
keras_tuner
pydub
ffmpeg
librosa
streamlit
yt-dlp